+++ bin/actions/0220022-12.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+#: BL+': Output name of foccussed application
+
+focusApplication=$(cat /proc/$(xdotool getwindowpid $(xdotool getwindowfocus))/comm)
+
+echo
+echo $focusApplication